requirements.txt
python-dotenv==0.21.0
aiogram==2.24
aiosqlite==0.18.0